#60 - refactor : ES 쿼리 코드 병합작업
ES 쿼리 관련 코드를 병합하는 작업 수행
- QueryBuilder2 -> QueryBuilder
- TransactionAggregationRepository2 -> TransactionAggregationRepository
- AnalysisService2 -> AnalysisService
- AnalysisController2 -> AnalysisController
로 병합 후 2번 객체들 모두 삭제
