+++ generator-api/src/main/java/com/simpaylog/generatorapi/repository/Elasticsearch/TransactionAggregationRepository.java
@@ -3,6 +3,7 @@ package com.simpaylog.generatorapi.repository.Elasticsearch;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.simpaylog.generatorapi.dto.analysis.AggregationInterval;
+import com.simpaylog.generatorapi.dto.analysis.HourlyTransaction;
 import com.simpaylog.generatorapi.dto.analysis.PeriodTransaction;
 import com.simpaylog.generatorapi.dto.analysis.TimeHeatmapCell;
 import com.simpaylog.generatorapi.utils.QueryBuilder;
@@ -15,10 +16,7 @@ import org.springframework.stereotype.Repository;
 
 import java.io.IOException;
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @Repository
 @RequiredArgsConstructor
@@ -79,6 +77,53 @@ public class TransactionAggregationRepository {
             }
         }
         return new TimeHeatmapCell(results);
+    }
+
+    public HourlyTransaction searchByHour(String sessionId, LocalDate from, LocalDate to) throws IOException {
+        Request request = new Request("GET", ES_END_POINT);
+        String queryJson = QueryBuilder.hourAggregationQuery(sessionId, from, to);
+        request.setJsonEntity(queryJson);
+
+        Response response = elasticsearchRestClient.performRequest(request);
+        String jsonResult = EntityUtils.toString(response.getEntity());
+
+        ObjectMapper objectMapper = new ObjectMapper();
+        JsonNode root = objectMapper.readTree(jsonResult);
+        JsonNode buckets = root.path("aggregations").path("by_transaction_type").path("buckets");
+
+        // 시간별 데이터 맵핑: hour -> HourlySummaryBuilder
+        Map<Integer, HourlyTransaction.HourlySummary> hourlyMap = new HashMap<>();
+
+        for (JsonNode typeBucket : buckets) {
+            String transactionType = typeBucket.path("key").asText();
+            JsonNode hourBuckets = typeBucket.path("by_hour").path("buckets");
+
+            for (JsonNode hourBucket : hourBuckets) {
+                int hour = hourBucket.path("key").asInt();
+                int count = hourBucket.path("transaction_count").path("value").asInt(0);
+                int avgAmount = hourBucket.path("average_amount").path("value").asInt(0);
+
+                HourlyTransaction.HourlySummary existing = hourlyMap.getOrDefault(hour, new HourlyTransaction.HourlySummary(hour, 0, 0, 0, 0));
+
+                HourlyTransaction.HourlySummary updated;
+                if ("WITHDRAW".equalsIgnoreCase(transactionType)) {
+                    updated = new HourlyTransaction.HourlySummary(hour, count, avgAmount, existing.totalIncomeCount(), existing.avgIncomeAmount());
+                } else {
+                    updated = new HourlyTransaction.HourlySummary(hour, existing.totalSpentCount(), existing.avgSpentAmount(), count, avgAmount);
+                }
+
+                hourlyMap.put(hour, updated);
+            }
+        }
+
+        // 0~23시간까지 빠진 시간도 포함해서 정렬
+        List<HourlyTransaction.HourlySummary> summaries = new ArrayList<>();
+        for (int i = 0; i < 24; i++) {
+            HourlyTransaction.HourlySummary hourlySummary = hourlyMap.getOrDefault(i, new HourlyTransaction.HourlySummary(0, 0, 0, 0, 0));
+            summaries.add(hourlySummary);
+        }
+
+        return new HourlyTransaction(summaries);
     }
 
     @SuppressWarnings("unchecked")
